examples: update react-router example to use latest libraries (#708)
