--- examples/_old/CONTRIBUTING.md
@@ -1,7 +1,0 @@
-Contributing Code to `flux-examples`
-------------------------------------
-
-Please be sure to sign our [CLA][] before you submit pull requests or otherwise contribute to `flux-examples`. This protects developers, who rely on [BSD license][].
-
-[BSD license]: https://github.com/yahoo/fluxible/blob/master/LICENSE.md
-[CLA]: https://yahoocla.herokuapp.com/
