--- examples/_old/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "fluxible-examples",
-  "private": true,
-  "version": "0.0.0",
-  "description": "Example applications using Fluxible",
-  "author": "Michael Ridgway <mridgway@yahoo-inc.com>",
-  "dependencies": {
-    "body-parser": "^1.6.4",
-    "classnames": "^2.1.3",
-    "cookie-parser": "^1.3.3",
-    "csurf": "^1.6.3",
-    "debug": "^2.0.0",
-    "express": "^4.12.3",
-    "express-state": "^1.2.0",
-    "fluxible": "^1.0.0",
-    "fluxible-addons-react": "^0.2.0",
-    "fluxible-plugin-fetchr": "^0.3.3",
-    "fluxible-router": "^1.0.0",
-    "history": "^1.17.0",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-router": "^3.0.0",
-    "serialize-javascript": "^4.0.0",
-    "serve-favicon": "^2.1.6"
-  },
-  "devDependencies": {
-    "babel-loader": "^6.0.0",
-    "grunt": "^1.0.1",
-    "grunt-cli": "^1.2.0",
-    "grunt-concurrent": "^2.3.1",
-    "grunt-contrib-clean": "^1.0.0",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-nodemon": "^0.4.0",
-    "grunt-webpack": "^1.0.8",
-    "nodemon": "^1.2.1",
-    "react-hot-loader": "^1.2.8",
-    "webpack": "^1.4.12",
-    "webpack-dev-server": "^1.6.5"
-  }
-}
