--- examples/_old/react-router/README.md
@@ -1,15 +1,0 @@
-# routing with react-router
-
-This shows routing on the client and the server with [react-router](https://github.com/rackt/react-router). Navigation uses history pushes on the client but browser refresh will render the current page correctly on the server.
-
-## Usage
-
-Make sure you run `npm install` from the root of flux-examples. Then change directories to routing (`cd routing`) and run:
-
-```bash
-$ npm run dev
-```
-
-This will use `nodemon` and `webpack` to watch for changes and restart and rebuild as needed.
-
-Open http://localhost:3000
