--- examples/_old/react-router/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "flux-example-routing",
-  "private": true,
-  "version": "0.0.0",
-  "description": "Example application",
-  "main": "start.js",
-  "author": "Michael Ridgway <mridgway@yahoo-inc.com>",
-  "scripts": {
-    "dev": "node webpack-dev-server.js & PORT=3001 ../node_modules/.bin/nodemon start.js -e js,jsx",
-    "start": "node start.js"
-  }
-}
