--- examples/_old/react-router/webpack.config.js
@@ -1,39 +1,0 @@
-var path = require('path');
-var webpack = require('webpack');
-
-module.exports = {
-    resolve: {
-        extensions: ['', '.js', '.jsx']
-    },
-    entry: [
-        'webpack-dev-server/client?http://localhost:3000',
-        'webpack/hot/only-dev-server',
-        './client.js'
-    ],
-    output: {
-        path: path.resolve('./build/js'),
-        filename: 'client.js',
-        publicPath: '/public/js/'
-    },
-    module: {
-        loaders: [
-            { test: /\.css$/, loader: 'style!css' },
-            {
-                test: /\.(js|jsx)$/,
-                exclude: /node_modules/,
-                loaders: [
-                    require.resolve('react-hot-loader'),
-                    require.resolve('babel-loader')
-                ]
-            }
-        ]
-    },
-    plugins: [
-        new webpack.HotModuleReplacementPlugin(),
-        new webpack.NoErrorsPlugin(),
-        // Protects against multiple React installs when npm linking
-        new webpack.NormalModuleReplacementPlugin(/^react?$/, require.resolve('react')),
-        new webpack.NormalModuleReplacementPlugin(/^react(\/addons)?$/, require.resolve('react/addons'))
-    ],
-    devtool: 'eval'
-};
