+++ examples/react-router/README.md
@@ -1,0 +1,18 @@
+# routing with react-router
+
+This shows routing on the client and the server with
+[react-router](https://github.com/rackt/react-router). Navigation uses
+history pushes on the client but browser refresh will render the
+current page correctly on the server.
+
+## Usage
+
+```bash
+npm install
+npm run dev
+```
+
+This will use `webpack` to watch for changes and restart and rebuild
+as needed.
+
+Open http://localhost:3000
