+++ examples/react-router/actions/navigate.js
