+++ examples/react-router/actions/updateTime.js
