+++ examples/react-router/client.js
@@ -4,28 +4,27 @@
  */
 /*global App, document, window */
 'use strict';
-var React = require('react');
-var ReactDOM = require('react-dom');
-var debug = require('debug');
-var bootstrapDebug = debug('Example');
-var app = require('./app');
-var dehydratedState = window.App; // Sent from the server
-var ReactRouter = require('react-router');
-var FluxibleComponent = require('fluxible-addons-react/FluxibleComponent');
-var createElement = require('fluxible-addons-react/createElementWithContext');
+const React = require('react');
+const ReactDOM = require('react-dom');
+const debug = require('debug');
+const bootstrapDebug = debug('Example');
+const app = require('./app').default;
+const dehydratedState = window.App; // Sent from the server
+const ReactRouter = require('react-router');
+const { FluxibleComponent } = require('fluxible-addons-react');
 
 bootstrapDebug('rehydrating app');
 
-function RenderApp(context){
+function RenderApp(context) {
     bootstrapDebug('React Rendering');
-    var mountNode = document.getElementById('app');
+    const mountNode = document.getElementById('app');
     ReactDOM.render(
         React.createElement(
             FluxibleComponent,
             { context: context.getComponentContext() },
             React.createElement(ReactRouter.Router, {
                 routes: context.getComponent(),
-                history: ReactRouter.browserHistory
+                history: ReactRouter.browserHistory,
             })
         ),
         mountNode,
