+++ examples/react-router/components/Application.js
@@ -5,22 +5,8 @@
 import React from 'react';
 import Nav from './Nav';
 import Timestamp from './Timestamp';
-import ApplicationStore from '../stores/ApplicationStore';
-import {connectToStores, provideContext} from 'fluxible-addons-react';
-import {RouteHandler, Router} from 'react-router';
-import routes from './Routes';
-import createBrowserHistory from 'history/lib/createBrowserHistory';
 
 class Application extends React.Component {
-
-    static contextTypes = {
-        getStore: React.PropTypes.func,
-        executeAction: React.PropTypes.func
-    };
-
-    constructor(props, context) {
-        super(props, context);
-    }
     render() {
         return (
             <div>
