+++ examples/react-router/components/Page.js
