+++ examples/react-router/stores/ApplicationStore.js
