+++ examples/react-router/stores/PageStore.js
