+++ examples/react-router/stores/TimeStore.js
